Move the equity function to the main model
